+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The changelog is release-oriented and uses a small fixed set of categories:
 
 ## [Unreleased]
 
+### Added
+- Typed loading of shared ped expression-set metadata.
+
 ## [0.3.13] - 2026-08-11
 
 ### Fixed